requirements.txt
# Core dependencies for profanity filtering
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0

# Note: PyTorch CPU-only installation
# For CPU-only: pip install torch torchaudio --index-url https://download.pytorch.org/whl/cpu

